third_party/WebKit/LayoutTests/http/tests/devtools/startup/sources/debugger/script-formatter-console-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that the script formatting changes console line numbers.


Running: testSetup

Running: testConsoleMessagesForFormattedScripts
script-formatter-console.html:9 
Pre-format row message list:
[8]
script-formatter-con…e.html:formatted:12 
Post-format row message list:
[11]
